requirements.txt
absl-py==2.1.0
adrf==0.1.4
asgiref==3.7.2
astunparse==1.6.3
async-property==0.2.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==3.0.4
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
daphne==3.0.2
Django==5.0.3
django-cors-headers==4.3.1
django-filter==23.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2024.2.0
gast==0.4.0
google-auth==2.28.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.0
h11==0.14.0
h5py==3.10.0
hyperlink==21.0.0
idna==3.6
imageio==2.34.0
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
keras==2.15.0
lazy_loader==0.3
libclang==16.0.6
load-dotenv==0.1.0
Markdown==3.5.2
MarkupSafe==2.1.5
migrate==0.3.8
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.8
networkx==3.2.1
numpy==1.24.3
oauthlib==3.2.2
openapi-codec==1.3.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pillow==10.2.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scipy==1.12.0
service-identity==24.1.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
swagger-ui-py==23.9.23
sympy==1.12
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tifffile==2024.2.12
torch==2.2.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.27.1
websockets==12.0
Werkzeug==3.0.1
wrapt==1.14.1
zope.interface==6.2
torchvision==0.18.0
ultralytics==8.2.28
scikit-learn==1.5.0